added an controller to handle the authentication related logic.
