+++ src/index.ts
@@ -9,6 +9,7 @@ const app = express();
 
 app.use(cors());
 app.use(express.json());
+
 app.use('/auth', authRoutes);
 app.use('/user', userRoutes);
 app.use('/tweet', authenticateToken, tweetRoutes);
@@ -17,9 +18,6 @@ app.get('/', (req, res) => {
     res.send("Hello World");
 });
 
-
-app.listen(3000, () => {
+app.listen(process.env.PORT || 3000, () => {
     console.log("Server ready at localhost:3000");
 });
-
-// I've stoped at 01:46:16 <------------
